fix(party): move voice phrase map out of ControllerView
Keep react-refresh happy by exporting mapTranscriptToKey from voicePhrases.ts.
