+++ apps/web/src/mp/ControllerView.tsx
@@ -3,20 +3,12 @@ import { useTranslation } from 'react-i18next';
 import { RoomClient, type RoomStatus } from './roomClient.js';
 import { useScreenWakeLock } from '../useScreenWakeLock.js';
 import type { InputKey, RoomPhase } from './protocol.js';
+import { mapTranscriptToKey } from './voicePhrases.js';
 
 type ControllerViewProps = { code: string; token: string };
 
 const NICK_ADJECTIVES = ['Swift', 'Brave', 'Sly', 'Wild', 'Calm', 'Bold', 'Lucky', 'Sneaky'];
 const NICK_ANIMALS = ['Fox', 'Otter', 'Lynx', 'Crane', 'Bison', 'Moth', 'Heron', 'Wolf'];
-
-/** Bilingual phrase → party key. One-shot recognition maps a final transcript onto these. */
-const VOICE_PHRASES: Array<{ key: InputKey; phrases: string[] }> = [
-  { key: 'up', phrases: ['up', 'góra', 'gora'] },
-  { key: 'down', phrases: ['down', 'dół', 'dol'] },
-  { key: 'left', phrases: ['left', 'lewo'] },
-  { key: 'right', phrases: ['right', 'prawo'] },
-  { key: 'a', phrases: ['a', 'go', 'fire', 'action', 'jump', 'akcja', 'strzał', 'strzal', 'ogień', 'ogien', 'skok'] },
-];
 
 interface SpeechRecognitionResultItem {
   transcript: string;
@@ -47,19 +39,6 @@ function randomNick(): string {
   const adjective = NICK_ADJECTIVES[Math.floor(Math.random() * NICK_ADJECTIVES.length)];
   const animal = NICK_ANIMALS[Math.floor(Math.random() * NICK_ANIMALS.length)];
   return `${adjective} ${animal}`;
-}
-
-function mapTranscriptToKey(transcript: string): InputKey | null {
-  const normalized = transcript.trim().toLowerCase().normalize('NFD').replace(/\p{M}/gu, '');
-  if (!normalized) return null;
-  const tokens = normalized.split(/[^a-z0-9]+/).filter(Boolean);
-  for (const entry of VOICE_PHRASES) {
-    for (const phrase of entry.phrases) {
-      const plain = phrase.normalize('NFD').replace(/\p{M}/gu, '');
-      if (tokens.includes(plain) || normalized === plain) return entry.key;
-    }
-  }
-  return null;
 }
 
 function speechRecognitionClass(): (new () => SpeechRecognitionInstance) | null {
@@ -335,6 +314,3 @@ export function ControllerView({ code, token }: ControllerViewProps) {
     </div>
   );
 }
-
-/** Exported for unit tests. */
-export { mapTranscriptToKey };
